Pipeline-operator-based solution in Haskell
